chore: rename and moved components
