+++ frontend/src/components/features/reactErrorBoundary/GlobalErrorBoundary.jsx
